split test / add error html (#91)
* add error html

* add api interface
