+++ test/file/search-gaokao-gzh-error.html
@@ -1,0 +1,104 @@
+
+<!DOCTYPE HTML>
+<html>
+<head>
+    <meta charset="utf-8">
+    <link rel="shortcut icon" href="//www.sogou.com/images/logo2014/new/favicon.ico" type="image/x-icon">
+    <title>搜狗搜索</title>
+    <link rel="stylesheet" href="static/css/anti.min.css?v=1"/>
+    <script src="//dl.web.sogoucdn.com/common/lib/jquery/jquery-1.11.0.min.js"></script>
+    <script src="static/js/antispider.min.js?v=2"></script>
+    <script>
+        var domain = getDomain();
+        window.imgCode = -1;
+
+        (function() {
+            function checkSNUID() {
+                var cookieArr = document.cookie.split('; '),
+                    count = 0;
+
+                for(var i = 0, len = cookieArr.length; i < len; i++) {
+                    if (cookieArr[i].indexOf('SNUID=') > -1) {
+                        count++;
+                    }
+                }
+
+                return count > 1;
+            }
+
+            if(checkSNUID()) {
+                var date = new Date(), expires;
+                date.setTime(date.getTime() -100000);
+
+                expires = date.toGMTString();
+
+                document.cookie = 'SNUID=1;path=/;expires=' + expires;
+                document.cookie = 'SNUID=1;path=/;expires=' + expires + ';domain=.www.sogou.com';
+                document.cookie = 'SNUID=1;path=/;expires=' + expires + ';domain=.weixin.sogou.com';
+                document.cookie = 'SNUID=1;path=/;expires=' + expires + ';domain=.sogou.com';
+                document.cookie = 'SNUID=1;path=/;expires=' + expires + ';domain=.snapshot.sogoucdn.com';
+
+                sendLog('delSNUID');
+            }
+
+            if(getCookie('seccodeRight') === 'success') {
+                sendLog('verifyLoop');
+
+                setCookie('seccodeRight', 1, getUTCString(-1), location.hostname, '/');
+            }
+
+            if(getCookie('refresh')) {
+                sendLog('refresh');
+            }
+        })();
+
+        function setImgCode(code) {
+            try {
+                var t = new Date().getTime() - imgRequestTime.getTime();
+                sendLog('imgCost',"cost="+t);
+            } catch (e) {
+            }
+            window.imgCode = code;
+        }
+        sendLog('index');
+
+        function changeImg2() {
+        	if(window.event) {
+        		window.event.returnValue=false
+        	}
+        }
+    </script>
+</head>
+<body>
+<div class="header">
+    <div class="logo"><a href="/"><img width="180" height="60" src="//www.sogou.com/images/logo2014/error180x60.png"></a></div>
+    <div class="other"><span class="s1">您的访问出错了</span><span class="s2"><a href="/">返回首页&gt;&gt;</a></span></div>
+</div>
+<div class="content-box">
+    <p class="ip-time-p">IP:123.116.247.15<br>访问时间：2017.07.25 22:36:19</p>
+    <p class="p2">用户您好，您的访问过于频繁，为确认本次访问为正常用户行为，需要您协助验证。</p>
+    <p class="p3"><label for="seccodeInput">验证码：</label></p>
+    <form name="authform" method="POST" id="seccodeForm" action="/">
+        <p class="p4">
+            <input type=text name="c" value="" placeholder="请输入验证码" id="seccodeInput">
+            <input type="hidden" name="tc" id="tc" value="">
+            <input type="hidden" name="r" id="from" value="%2Fweixin%3Ftype%3D2%26query%3D%E9%AB%98%E8%80%83%26ie%3Dutf8%26s_from%3Dinput%26_sug_%3Dn%26_sug_type_%3D1%26w%3D01015002%26oq%3D%26ri%3D6%26sourceid%3Dsugg%26sut%3D0%26sst0%3D1500993227871%26lkt%3D0%2C0%2C0%26p%3D40040108" >
+            <input type="hidden" name="m" value="0" >            <span class="s1">
+                <script>imgRequestTime=new Date();</script>
+                <a onclick="changeImg2();" href="javascript:void(0)">
+                    <img id="seccodeImage" onload="setImgCode(1)" onerror="setImgCode(0)" src="util/seccode.php?tc=1500993379" width="100" height="40" alt="请输入图中的验证码" title="请输入图中的验证码">
+                </a>
+            </span>
+            <a href="javascript:void(0);" id="change-img" onclick="changeImg2();" style="padding-left:50px;">换一张</a>
+            <span class="s2" id="error-tips" style="display: none;"></span>
+        </p>
+    </form>
+    <p class="p5">
+        <a href="javascript:void(0);" id="submit">提交</a>
+        <span>提交后没解决问题？欢迎<a href="http://fankui.help.sogou.com/index.php/web/web/index?type=10&anti_time=1500993379&domain=weixin.sogou.com" target="_blank">反馈</a>。</span>
+    </p>
+</div>
+<div id="ft"><a href="http://fuwu.sogou.com/" target="_blank">企业推广</a><a href="http://corp.sogou.com/" target="_blank">关于搜狗</a><a href="/docs/terms.htm?v=1" target="_blank">免责声明</a><a href="http://fankui.help.sogou.com/index.php/web/web/index?type=10&anti_time=1500993379&domain=weixin.sogou.com" target="_blank">意见反馈</a><br>&nbsp;&copy;&nbsp;2017<span id="footer-year"></span>&nbsp;SOGOU&nbsp;-&nbsp;<a href="http://www.miibeian.gov.cn" target="_blank" class="g">京ICP证050897号</a>&nbsp;-&nbsp;京公网安备1100<span class="ba">00000025号</span></div>
+<script src="static/js/index.min.js?v=0.1.3"></script>
+</body>
+</html><!--zly-->
