+++ test/test_api.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+import unittest
+
+
+class TestAPI(unittest.TestCase):
+    @unittest.skip
+    def test_api(self):
+        pass
+
+
+if __name__ == '__main__':
+    unittest.main()
