+++ test/test_main.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import, unicode_literals, print_function
+
+from wechatsogou.refactor_request import WechatSogouRequest
+from wechatsogou.refactor_structuring import WechatSogouStructuring
+
+ws = WechatSogouRequest()
+ws_structuring = WechatSogouStructuring()
+
+gaokao_keyword = '高考'
+fake_data_path = 'test/file'
