+++ wechatsogou/refactor_api.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import absolute_import, unicode_literals, print_function
+
+
+class WechatSogouAPI(object):
+    pass
